libc/stdio: legacy vfprintf uses wcsrtombs with wrong types
Need to cast the return to (int) to make it compatible with other code
in this file.

+++ newlib/libc/stdio/vfprintf.c
@@ -1483,7 +1483,7 @@ string:
 
 				/* Convert widechar string to multibyte string. */
 				memset ((void *)&ps, '\0', sizeof (mbstate_t));
-				if (wcsrtombs (cp, &wcp, size, &ps)
+				if ((int)wcsrtombs (cp, &wcp, size, &ps)
 				    != size) {
 					fp->_flags |= __SERR;
 					goto error;
